refactor: move cloudflare singleton
