+++ app/server/cloudflare-client.ts
@@ -1,0 +1,5 @@
+import Cloudflare from "cloudflare";
+
+export const cloudflareClient = new Cloudflare({
+  apiKey: process.env.CLOUDFLARE_API_TOKEN,
+});
